Image_deblurring/requirements.txt
torch==2.2.0
torchvision==0.17.0
numpy==1.24.3
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.12.0
matplotlib==3.8.3
Pillow==10.2.0
tqdm==4.66.2
einops==0.7.0
lpips==0.1.4
openvino==2023.1.0
